fix: laravel 11 migration syntax
In laravel 11, they changed the syntax for changing columns so that all constraints are removed and reapplied, compared to only adding constraints that weren't added before. I generated the sql code of each migration on laravel 10, and the same on laravel 11, and compared the sql code between the two versions and changed the migrations so that the sql matched between the two versions.

+++ database/migrations/2017_12_10_020753_create_request_bounty_table.php
@@ -25,7 +25,7 @@ return new class () extends Migration {
         Schema::create('request_bounty', function (Blueprint $table): void {
             $table->integer('id', true);
             $table->integer('user_id')->index('addedby');
-            $table->float('seedbonus', 12)->unsigned()->default(0.00);
+            $table->double('seedbonus')->unsigned()->default(0.00);
             $table->integer('requests_id')->index('request_id');
             $table->timestamps();
         });
