+++ database/migrations/2017_12_10_020753_create_requests_table.php
@@ -33,7 +33,7 @@ return new class () extends Migration {
             $table->string('mal')->nullable()->index('mal');
             $table->text('description');
             $table->integer('user_id')->index('requests_user_id_foreign');
-            $table->float('bounty', 22);
+            $table->double('bounty');
             $table->integer('votes')->default(0);
             $table->boolean('claimed')->nullable();
             $table->timestamps();
